homepage: remove old and split out new
