+++ hyperdrive/packages/homepage/ui/src/components/AndroidHomescreen/styles/animations.css
@@ -1,0 +1,8 @@
+@keyframes wiggle {
+  0%, 100% { transform: rotate(-3deg); }
+  50% { transform: rotate(3deg); }
+}
+
+.animate-wiggle {
+  animation: wiggle 0.3s ease-in-out infinite;
+}
